Fix a warning with POPCNT and MSVC
Intrinsic __popcnt64() returns an unsigned __int64, cast
to an integer and silence the warning.

No functional change.

+++ src/bitcount.h
@@ -51,7 +51,7 @@ inline bool cpu_has_popcnt() {
   return (CPUInfo[2] >> 23) & 1;
 }
 
-#define POPCNT_INTRINSIC(x) __popcnt64(x)
+#define POPCNT_INTRINSIC(x) (int)__popcnt64(x)
 
 #elif defined(__GNUC__) && defined(USE_POPCNT) // Gcc compiler
